% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi_wrapper.R
\name{binary_multi_evppi}
\alias{binary_multi_evppi}
\title{Calculate Expected Value of Perfect Parameter Information (EVPPI) for
multiple input variables and two decision options (yes/no).}
\usage{
binary_multi_evppi(x, y)
}
\arguments{
\item{x}{Monte Carlo samples from the probability distribution of the
considered parameter (aka estimates aka "input" variables). Columns are
variables, rows are samples.}

\item{y}{The respective utility (aka outcome) samples calculated using the
estimate samples above. This utility is considered to be the (only)
decision criterion for a risk-neutral decision maker facing a binary
decision, so that a positive expected value will lead to `yes` and a
negative one to `no`.}
}
\value{
Vector of EVPPI values in the order of the columns of x.
}
\description{
Calculate Expected Value of Perfect Parameter Information (EVPPI) for
multiple input variables and two decision options (yes/no).
}
