% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evpi_wrapper.R
\name{multi_evppi}
\alias{multi_evppi}
\title{Calculate Expected Value of Perfect Parameter Information (EVPPI) for
multiple input variables and multiple decision options.}
\usage{
multi_evppi(x, y)
}
\arguments{
\item{x}{Monte Carlo samples from the probability distribution of the
considered parameter (aka estimates aka "input" variables). Columns are
variables, rows are samples.}

\item{y}{The respective utility (aka outcome) samples calculated using the
estimate samples of x. This criterion is considered to be the (only)
decision criterion for a risk-neutral decision maker, that chooses
the option with the highest expected utility. Samples are rows,
decision options are columns.}
}
\value{
Vector of EVPPI values in the order of the columns of `x`.
}
\description{
Calculate Expected Value of Perfect Parameter Information (EVPPI) for
multiple input variables and multiple decision options.
}
